# docs/swagger.yaml
basePath: /api
definitions:
  ds.Letter:
    properties:
      description:
        type: string
      details:
        type: string
      id:
        type: integer
      imageURL:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      periodEnd:
        type: integer
      periodStart:
        type: integer
    type: object
  ds.Manuscript:
    properties:
      calculatedPeriod:
        type: string
      createdAt:
        type: string
      finishedAt:
        type: string
      id:
        type: integer
      letters:
        items:
          $ref: '#/definitions/ds.ManuscriptLetter'
        type: array
      manuscript_text:
        type: string
      moderator:
        $ref: '#/definitions/ds.User'
      moderatorID:
        type: integer
      status:
        type: string
      submittedAt:
        type: string
      user:
        $ref: '#/definitions/ds.User'
      user_id:
        type: integer
    type: object
  ds.ManuscriptLetter:
    properties:
      letter:
        $ref: '#/definitions/ds.Letter'
      letterID:
        type: integer
      manuscript:
        $ref: '#/definitions/ds.Manuscript'
      manuscriptID:
        type: integer
      quantity:
        type: integer
    type: object
  ds.ManuscriptListItemLetterResponse:
    properties:
      description:
        type: string
      letter_id:
        type: integer
      name:
        type: string
      period_end:
        type: integer
      period_start:
        type: integer
      quantity:
        description: 'Важно: количество из связующей таблицы ManuscriptLetter'
        type: integer
    type: object
  ds.ManuscriptListItemResponse:
    properties:
      calculated_period:
        type: string
      id:
        type: integer
      letters:
        description: Вложенные письма
        items:
          $ref: '#/definitions/ds.ManuscriptListItemLetterResponse'
        type: array
      manuscript_text:
        description: 'ПРИМЕЧАНИЕ: Если поле manuscript_text существует'
        type: string
      moderator_id:
        description: Информация о Модераторе (опционально)
        type: integer
      moderator_name:
        description: Будет nil, если модератор не назначен
        type: string
      status:
        type: string
      user_id:
        description: Информация о Создателе
        type: integer
      username:
        type: string
    type: object
  ds.User:
    properties:
      id:
        type: integer
      password:
        type: string
      role:
        $ref: '#/definitions/role.Role'
      username:
        type: string
    type: object
  handler.BasketStatusResponse:
    properties:
      letter_count:
        example: 5
        type: integer
      manuscript_id:
        example: 10
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: invalid json or record not found
        type: string
    type: object
  handler.ImageUploadResponse:
    properties:
      image_url:
        example: http://127.0.0.1:9000/manuscripts/10_uuid.jpg
        type: string
      status:
        example: Image successfully uploaded and link updated
        type: string
    type: object
  handler.LoginRequest:
    properties:
      password:
        example: testpassword
        type: string
      username:
        example: testuser
        type: string
    type: object
  handler.LoginResponse:
    properties:
      role:
        example: Buyer
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  handler.ModerateRequest:
    properties:
      status:
        description: Должен быть 'finished' или 'rejected'
        example: finished
        type: string
    type: object
  handler.RegisterResponse:
    properties:
      user_id:
        example: 42
        type: integer
    type: object
  handler.StatusResponse:
    properties:
      status:
        example: updated
        type: string
    type: object
  handler.UpdateQuantityRequest:
    properties:
      quantity:
        example: 3
        type: integer
    type: object
  handler.UpdateUserInput:
    properties:
      password:
        example: newsecurepassword
        type: string
      username:
        example: updateduser
        type: string
    type: object
  handler.UserRegisterInput:
    properties:
      password:
        example: securepassword
        type: string
      username:
        example: newuser
        type: string
    type: object
  role.Role:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      Admin: 2 - Администратор
      Buyer: 0 - Обычный пользователь
      Manager: 1 - Модератор
    x-enum-descriptions:
    - 0 - Обычный пользователь
    - 1 - Модератор
    - 2 - Администратор
    x-enum-varnames:
    - Buyer
    - Manager
    - Admin
host: localhost:8081
info:
  contact:
    email: support@example.com
    name: API Support
  description: API для работы с рукописями, письмами и пользователями.
  license:
    name: AS IS (NO WARRANTY)
  title: Manuscript API
  version: "1.0"
paths:
  /letters:
    get:
      description: Возвращает список всех активных писем, опционально фильтруя по
        тексту. Доступно всем.
      parameters:
      - description: Фильтр для поиска по имени/описанию
        in: query
        name: filter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список писем
          schema:
            items:
              $ref: '#/definitions/ds.Letter'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить список писем (Letter)
      tags:
      - Letters
    post:
      consumes:
      - application/json
      description: Создает новое письмо. Требуется роль Manager или Admin.
      parameters:
      - description: Данные нового письма
        in: body
        name: letter
        required: true
        schema:
          $ref: '#/definitions/ds.Letter'
      produces:
      - application/json
      responses:
        "201":
          description: Созданное письмо
          schema:
            $ref: '#/definitions/ds.Letter'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создать новое письмо
      tags:
      - Letters
  /letters/{id}:
    delete:
      description: Удаляет письмо по ID. Требуется роль Manager или Admin.
      parameters:
      - description: ID письма
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить письмо
      tags:
      - Letters
    get:
      description: Возвращает письмо по его ID. Доступно всем.
      parameters:
      - description: ID письма
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Детали письма
          schema:
            $ref: '#/definitions/ds.Letter'
        "404":
          description: Письмо не найдено
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить конкретное письмо
      tags:
      - Letters
    put:
      consumes:
      - application/json
      description: Обновляет данные письма по ID. Требуется роль Manager или Admin.
      parameters:
      - description: ID письма
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные письма
        in: body
        name: letter
        required: true
        schema:
          $ref: '#/definitions/ds.Letter'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить письмо
      tags:
      - Letters
  /letters/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение в MinIO и сохраняет URL в базе данных для
        указанного письма. Требуется роль Manager или Admin.
      parameters:
      - description: ID письма
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Успешная загрузка
          schema:
            $ref: '#/definitions/handler.ImageUploadResponse'
        "400":
          description: Неверный ID или файл не предоставлен
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера (MinIO или БД)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Загрузить изображение для письма
      tags:
      - Letters
  /letters/{id}/manuscript:
    post:
      consumes:
      - application/json
      description: Добавляет указанное письмо в текущий черновик рукописи пользователя.
        Если черновика нет, он создается. Требуется роль Buyer.
      parameters:
      - description: ID письма
        in: path
        name: id
        required: true
        type: integer
      - description: Количество
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateQuantityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Письмо добавлено
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Добавить письмо в черновик рукописи
      tags:
      - Letters
  /manuscripts:
    get:
      description: Возвращает список рукописей, отфильтрованный по статусу и/или датам.
        Администраторы и модераторы видят все заявки, покупатели – только свои.
      parameters:
      - description: Фильтр по статусу (например, 'submitted', 'finished')
        in: query
        name: status
        type: string
      - description: Начальная дата (YYYY-MM-DD)
        in: query
        name: start
        type: string
      - description: Конечная дата (YYYY-MM-DD)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список рукописей" // <-- ИСПРАВЛЕНО
          schema:
            items:
              $ref: '#/definitions/ds.ManuscriptListItemResponse'
            type: array
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Фильтрация списка рукописей
      tags:
      - Manuscripts
  /manuscripts/{id}:
    delete:
      description: Удаляет рукопись по ID. Доступно владельцу (Owner) в статусе 'draft'
        или Admin.
      parameters:
      - description: ID рукописи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить рукопись
      tags:
      - Manuscripts
    get:
      description: Возвращает детали рукописи по ID. Требуется авторизация (Owner,
        Manager, Admin).
      parameters:
      - description: ID рукописи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Детали рукописи
          schema:
            $ref: '#/definitions/ds.Manuscript'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Рукопись не найдена
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить конкретную рукопись
      tags:
      - Manuscripts
    put:
      consumes:
      - application/json
      description: Обновляет поля рукописи по ID. Доступно только владельцу (Owner)
        и только в статусе 'draft'.
      parameters:
      - description: ID рукописи
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные рукописи
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/ds.Manuscript'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить рукопись
      tags:
      - Manuscripts
  /manuscripts/{id}/letters/{lid}:
    delete:
      description: Удаляет связь между письмом (Letter) и рукописью (Manuscript).
        Доступно владельцу (Owner) в статусе 'draft'.
      parameters:
      - description: ID рукописи
        in: path
        name: id
        required: true
        type: integer
      - description: ID письма
        in: path
        name: lid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить письмо из рукописи
      tags:
      - Manuscripts
    put:
      consumes:
      - application/json
      description: Обновляет количество конкретного письма (Letter) внутри рукописи
        (Manuscript). Доступно владельцу (Owner) в статусе 'draft'.
      parameters:
      - description: ID рукописи
        in: path
        name: id
        required: true
        type: integer
      - description: ID письма
        in: path
        name: lid
        required: true
        type: integer
      - description: Новое количество
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateQuantityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить количество письма в рукописи
      tags:
      - Manuscripts
  /manuscripts/{id}/moderation:
    put:
      consumes:
      - application/json
      description: Завершает ('finished') или отклоняет ('rejected') рукопись. Требуется
        роль Manager или Admin.
      parameters:
      - description: ID рукописи
        in: path
        name: id
        required: true
        type: integer
      - description: 'Статус модерации: ''finished'' или ''rejected'''
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.ModerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная модерация
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Неверный статус или ошибка БД
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Модерация рукописи
      tags:
      - Manuscripts
  /manuscripts/{id}/submit:
    put:
      description: Меняет статус рукописи с 'draft' на 'submitted'. Доступно только
        владельцу (Owner).
      parameters:
      - description: ID рукописи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешно отправлено
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Неверный статус или ошибка БД
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Отправить рукопись на модерацию
      tags:
      - Manuscripts
  /manuscripts/basket:
    get:
      description: Возвращает ID текущего черновика (Manuscript) пользователя и количество
        писем в нем. Требуется роль Buyer.
      produces:
      - application/json
      responses:
        "200":
          description: Статус черновика
          schema:
            $ref: '#/definitions/handler.BasketStatusResponse'
        "401":
          description: Неавторизован
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Нет доступа
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Черновик не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Статус корзины/черновика
      tags:
      - Manuscripts
  /users/{id}:
    get:
      description: 'Возвращает данные пользователя по его ID (заглушка: всегда ID
        1). Требуется авторизация.'
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/ds.User'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить данные пользователя
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: 'Обновляет имя и/или пароль пользователя (заглушка: всегда ID 1).
        Требуется авторизация.'
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные пользователя
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить данные пользователя
      tags:
      - Users
  /users/login:
    post:
      consumes:
      - application/json
      description: Вход в систему, проверка учетных данных и выдача JWT-токена.
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход и токен
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка генерации токена
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Аутентификация пользователя
      tags:
      - Users
  /users/logout:
    post:
      consumes:
      - application/json
      description: Добавляет текущий JWT-токен в черный список Redis. Требуется авторизация.
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
          schema:
            $ref: '#/definitions/handler.StatusResponse'
        "401":
          description: Токен не предоставлен/недействителен
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера (Redis)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Выход из системы
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя с ролью Buyer по умолчанию.
      parameters:
      - description: Данные нового пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.UserRegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/handler.RegisterResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера (например, пользователь уже существует)
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - Users
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"